web-demo/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9 # <--- 更改這裡：使用 psycopg2-binary